refactor: device 로 패키지구조화
