--- monicar-emulator/src/main/java/org/emulator/application/port/EmulatorRepository.java
@@ -1,11 +1,0 @@
-package org.emulator.application.port;
-
-import org.emulator.domain.VehicleInfo;
-
-/**
- * 차량 정보 저장 역할
- */
-public interface EmulatorRepository {
-    int getTotalDistance();
-	String getGpsStatus();
-}
