--- monicar-emulator/src/main/java/org/emulator/application/port/LocationEventPublisher.java
@@ -1,7 +1,0 @@
-package org.emulator.application.port;
-
-/**
- * 위치 정보 이벤트 발행 역할
- */
-public interface LocationEventPublisher {
-}
