--- monicar-emulator/src/main/java/org/emulator/application/port/LocationReceiver.java
@@ -1,7 +1,0 @@
-package org.emulator.application.port;
-
-/**
- * 임의의 파일 or GPS 센서로부터 gps 정보를 받아온다.
- */
-public interface LocationReceiver {
-}
