+++ monicar-emulator/src/main/java/org/emulator/device/application/port/EmulatorRepository.java
@@ -1,0 +1,9 @@
+package org.emulator.device.application.port;
+
+/**
+ * 차량 정보 저장 역할
+ */
+public interface EmulatorRepository {
+    int getTotalDistance();
+	String getGpsStatus();
+}
