+++ monicar-emulator/src/main/java/org/emulator/device/application/port/LocationEventPublisher.java
@@ -1,0 +1,7 @@
+package org.emulator.device.application.port;
+
+/**
+ * 위치 정보 이벤트 발행 역할
+ */
+public interface LocationEventPublisher {
+}
