+++ monicar-emulator/src/main/java/org/emulator/device/application/port/LocationReceiver.java
@@ -1,0 +1,7 @@
+package org.emulator.device.application.port;
+
+/**
+ * 임의의 파일 or GPS 센서로부터 gps 정보를 받아온다.
+ */
+public interface LocationReceiver {
+}
