+++ monicar-emulator/src/main/java/org/emulator/device/common/RequestUtils.java
@@ -1,0 +1,17 @@
+package org.emulator.device.common;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.UUID;
+
+public final class RequestUtils {
+	private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("yyyyMMddHHmmss");
+
+	public static String getCurrentTimestamp() {
+		return LocalDateTime.now().format(DATE_TIME_FORMATTER);
+	}
+
+	public static String generateTUID() {
+		return UUID.randomUUID().toString();
+	}
+}
