+++ monicar-emulator/src/main/java/org/emulator/device/common/VehicleConstant.java
@@ -1,0 +1,16 @@
+package org.emulator.device.common;
+
+public final class VehicleConstant {
+	public static final String VEHICLE_NUM = "01234567890";
+	public static final String TERMINAL_ID = "A001";
+	public static final String MANUFACTURER_ID = "6";
+	public static final String PACKET_VERSION = "5";
+	public static final String DEVICE_ID = "1";
+
+	public static final String CONTROL_CODE = "05";
+	public static final String DFW_VERSION = "LTE 1.2";
+
+	private VehicleConstant() {
+		throw new AssertionError();
+	}
+}
