+++ monicar-emulator/src/main/java/org/emulator/device/domain/GpsStatus.java
@@ -1,0 +1,5 @@
+package org.emulator.device.domain;
+
+public enum GpsStatus {
+	A;
+}
