+++ monicar-emulator/src/main/java/org/emulator/device/infrastructure/external/UrlPathEnum.java
@@ -1,0 +1,16 @@
+package org.emulator.device.infrastructure.external;
+
+public enum UrlPathEnum {
+	CONTROL_CENTER("control-center"),
+	COLLECTOR("collector");
+
+	private final String path;
+
+	UrlPathEnum(String path) {
+		this.path = path;
+	}
+
+	public String getApiUrl() {
+		return "http://localhost:8090/api/v1" + path + "/";
+	}
+}
