+++ monicar-emulator/src/main/java/org/emulator/device/infrastructure/file/FileLocationReceiver.java
@@ -1,0 +1,8 @@
+package org.emulator.device.infrastructure.file;
+
+import org.emulator.device.application.port.LocationReceiver;
+import org.springframework.stereotype.Component;
+
+@Component
+public class FileLocationReceiver implements LocationReceiver {
+}
