+++ monicar-emulator/src/main/java/org/emulator/device/infrastructure/messaging/KafkaLocationEventPublisher.java
@@ -1,0 +1,8 @@
+package org.emulator.device.infrastructure.messaging;
+
+import org.emulator.device.application.port.LocationEventPublisher;
+import org.springframework.stereotype.Component;
+
+@Component
+public class KafkaLocationEventPublisher implements LocationEventPublisher {
+}
