--- monicar-emulator/src/main/java/org/emulator/domain/ControlInfo.java
@@ -1,12 +1,0 @@
-package org.emulator.domain;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class ControlInfo {
-	private final String controlId;
-	private final String controlCode;
-	private final String cycleInfoTransmitTerm;
-}
