--- monicar-emulator/src/main/java/org/emulator/domain/GpsStatus.java
@@ -1,5 +1,0 @@
-package org.emulator.domain;
-
-public enum GpsStatus {
-	A;
-}
