--- monicar-emulator/src/main/java/org/emulator/infrastructure/external/RestClientFactory.java
@@ -1,35 +1,0 @@
-package org.emulator.infrastructure.external;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
-
-import org.emulator.common.HeaderUtils;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestClient;
-
-@Configuration
-public class RestClientFactory {
-
-	@Bean
-	public Map<UrlPathEnum, RestClient> restClients() {
-		Map<UrlPathEnum, RestClient> restClients = new HashMap<>();
-
-		Arrays.stream(UrlPathEnum.values()).forEach(apiUrl -> {
-			restClients.put(apiUrl, RestClient.builder()
-				.baseUrl(apiUrl.getApiUrl())
-				.defaultHeaders(HeaderUtils.defaultHeaders())
-				.build());
-		});
-
-		return restClients;
-	}
-
-	@Bean
-	public RestClientService restClientService(Map<UrlPathEnum, RestClient> restClients) {
-		return new RestClientService(restClients, new ObjectMapper());
-	}
-}
