--- monicar-emulator/src/main/java/org/emulator/infrastructure/external/UrlPathEnum.java
@@ -1,16 +1,0 @@
-package org.emulator.infrastructure.external;
-
-public enum UrlPathEnum {
-	CONTROL_CENTER("control-center"),
-	COLLECTOR("collector");
-
-	private final String path;
-
-	UrlPathEnum(String path) {
-		this.path = path;
-	}
-
-	public String getApiUrl() {
-		return "http://localhost:8090/api/v1" + path + "/";
-	}
-}
