--- monicar-emulator/src/main/java/org/emulator/infrastructure/file/FileLocationReceiver.java
@@ -1,8 +1,0 @@
-package org.emulator.infrastructure.file;
-
-import org.emulator.application.port.LocationReceiver;
-import org.springframework.stereotype.Component;
-
-@Component
-public class FileLocationReceiver implements LocationReceiver {
-}
