--- monicar-emulator/src/main/java/org/emulator/infrastructure/messaging/KafkaLocationEventPublisher.java
@@ -1,8 +1,0 @@
-package org.emulator.infrastructure.messaging;
-
-import org.emulator.application.port.LocationEventPublisher;
-import org.springframework.stereotype.Component;
-
-@Component
-public class KafkaLocationEventPublisher implements LocationEventPublisher {
-}
